Tries to use 64 bit integers for total score and ranked sscore.

+++ shared/database/entities/OsuDroidUser.ts
@@ -62,10 +62,10 @@ export default class OsuDroidUser
   @Column("float")
   pp!: number;
 
-  @Column("int")
+  @Column("int64")
   rankedScore!: number;
 
-  @Column("int")
+  @Column("int64")
   totalScore!: number;
 
   @Column("string", { array: true })
